PM-443 Moving market record to src/store/models
