+++ src/routes/marketlist/store/test/openMarkets.selector.js
@@ -1,6 +1,6 @@
 import moment from 'moment'
 import { List } from 'immutable'
-import { MARKET_STAGES } from '../models/market'
+import { MARKET_STAGES } from 'store/models'
 import { REDUCER_ID } from '../reducers/market'
 import { openMarketSelector } from '../selectors'
 import aMarket from './builder/index.builder'
